Move pan-view components to their own subdir.
This makes it easier to modularize pan-view without cluttering the src dir
with extra files, and without worrying as much about exposing pan-view-internal
APIs to the rest of Geeqie.
