+++ src/pan-view/Makefile.am
@@ -1,0 +1,16 @@
+module_pan_view = \
+	%D%/pan-calendar.c	\
+	%D%/pan-calendar.h	\
+	%D%/pan-folder.c	\
+	%D%/pan-folder.h	\
+	%D%/pan-grid.c	\
+	%D%/pan-grid.h	\
+	%D%/pan-item.c	\
+	%D%/pan-item.h	\
+	%D%/pan-timeline.c	\
+	%D%/pan-timeline.h	\
+	%D%/pan-types.h	\
+	%D%/pan-util.c	\
+	%D%/pan-util.h	\
+	%D%/pan-view.c	\
+	%D%/pan-view.h
