+++ src/pan-view/pan-calendar.c
