+++ src/pan-view/pan-folder.c
