+++ src/pan-view/pan-folder.h
@@ -19,12 +19,15 @@
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 
-#ifndef PAN_VIEW_H
-#define PAN_VIEW_H
+#ifndef PAN_VIEW_PAN_FOLDER_H
+#define PAN_VIEW_PAN_FOLDER_H
 
-#include "filedata.h"
+#include "main.h"
+#include "pan-types.h"
 
-void pan_window_new(FileData *dir_fd);
+void pan_flower_compute(PanWindow *pw, FileData *dir_fd,
+			gint *width, gint *height,
+			gint *scroll_x, gint *scroll_y);
+void pan_folder_tree_compute(PanWindow *pw, FileData *dir_fd, gint *width, gint *height);
 
 #endif
-/* vim: set shiftwidth=8 softtabstop=0 cindent cinoptions={1s: */
