+++ src/pan-view/pan-grid.c
