+++ src/pan-view/pan-item.c
