+++ src/pan-view/pan-timeline.c
