+++ src/pan-view/pan-util.c
