+++ src/pan-view/pan-view.c
